Module algorithmia::algo

Interact with Algorithmia algorithms

Instantiate from the Algorithmia struct

Structs

AlgoMetadata
Metadata returned from the API

AlgoOptions
Options used to alter the algorithm call, e.g. configuring the timeout

AlgoRef

AlgoResponse
Successful API response that wraps the AlgoIo and its Metadata

Algorithm
Algorithmia algorithm - intialized from the Algorithmia builder

Enums

AlgoIo
Types that represent either input to or output from an algorithm

Version
Version of an algorithm

Traits

DecodedEntryPoint
Alternate implementation for EntryPoint that automatically decodes JSON input to the associate type.

EntryPoint
Implementing an algorithm involves overriding at least one of these methods