Enum algorithmia::algo::AlgoInput

pub enum AlgoInput<'a> {
    Text(Cow<'a, str>),
    Binary(Cow<'a, [u8]>),
    Json(Cow<'a, Json>),
}

Types that can be used as input to an algorithm

Variants
Text(Cow<'a, str>)

Data that will be sent with Content-Type: text/plain

Binary(Cow<'a, [u8]>)

Data that will be sent with Content-Type: application/octet-stream

Json(Cow<'a, Json>)

Data that will be sent with Content-Type: application/json

Methods

impl<'a> AlgoInput<'a>

fn as_string(&'a self) -> Option<&'a str>

If the AlgoInput is text (or a valid JSON string), returns the associated text

fn as_json(&'a self) -> Option<Cow<'a, Json>>

If the AlgoInput is Json (or text that can be JSON encoded), returns the associated JSON string

For AlgoInput::Json, this returns the borrowed Json. For the AlgoInput::Text variant, the text is wrapped into an owned Json::String.

fn as_bytes(&'a self) -> Option<&'a [u8]>

If the AlgoInput is binary, returns the associated byte slice

fn decode<D: Decodable>(&self) -> Result<D, Error>

If the AlgoInput is valid JSON, decode it to a particular type