Enum algorithmia::algo::AlgoIo

pub enum AlgoIo {
    Text(String),
    Json(Json),
    Binary(Vec<u8>),
}

Types that represent either input to or output from an algorithm

Variants

Text

Input or output data that represents plain text (text/plain)

Json

Input or output text that represents JSON (application/json)

Binary

Input or output text that represents binary data (application/octet-stream)

Methods

impl AlgoIo

fn as_string<'a>(&'a self) -> Option<&'a str>

If the AlgoIo is text (or a valid JSON string), returns the associated text

fn as_json<'a>(&'a self) -> Option<Cow<'a, Json>>

If the AlgoIo is Json (or text that can be JSON encoded), returns the associated JSON string

fn as_bytes<'a>(&'a self) -> Option<&'a [u8]>

If the AlgoIo is binary, returns the associated byte slice

fn decode<D: Decodable>(&self) -> Result<D, Error>

If the AlgoIo is valid JSON, decode it to a particular type