Struct algorithmia::algo::Algorithm

pub struct Algorithm {
    pub path: String,
    // some fields omitted
}

Algorithmia algorithm

Fields

path

Methods

impl Algorithm

fn new(client: HttpClient, algo_ref: AlgoRef) -> Algorithm

fn to_url(&self) -> Url

Get the API Endpoint URL for this Algorithm

fn to_algo_uri(&self) -> String

Get the Algorithmia algo URI for this Algorithm

fn pipe<'a, I>(&'a self, input_data: I, options: Option<&'a AlgoOptions>) -> Result<AlgoResponse, Error> where I: Into<AlgoInput<'a>>

Execute an algorithm with

Content-type is determined by the type of input_data String => plain/text Encodable => application/json Byte slice => application/octet-stream

To create encodable objects for complex input, use #[derive(RustcEncodable)] on your struct

If you want a string to be sent as application/json, use pipe_json(...) instead

Examples

let client = Algorithmia::client("111112222233333444445555566");
let moving_avg = client.algo("timeseries/SimpleMovingAverage/0.1");
let input = (vec![0,1,2,3,15,4,5,6,7], 3);
match moving_avg.pipe(&input, None) {
    Ok(response) => println!("{}", response.result_json().unwrap()),
    Err(err) => println!("ERROR: {}", err),
};

fn pipe_json(&self, json_input: &str, options: Option<&AlgoOptions>) -> Result<AlgoResponse, Error>

Execute an algorithm with explicitly set content-type

pipe provides a JSON encoding/decoding wrapper around this method

Examples

let client = Algorithmia::client("111112222233333444445555566");
let minmax  = client.algo("codeb34v3r/FindMinMax/0.1");

let output = match minmax.pipe_json("[2,3,4]", None) {
   Ok(response) => response.result_json().unwrap().to_owned(),
   Err(err) => panic!("{}", err),
};

fn pipe_as<'a, B>(&'a self, input_data: B, content_type: Mime, options: Option<&'a AlgoOptions>) -> Result<Response, Error> where B: Into<Body<'a>>