Trait algorithmia::algo::EntryPoint

pub trait EntryPoint: Default {
    fn apply_str(&self, name: &str) -> Result<AlgoIo, Box<Error>> { ... }
    fn apply_json(&self, json: &Json) -> Result<AlgoIo, Box<Error>> { ... }
    fn apply_bytes(&self, bytes: &[u8]) -> Result<AlgoIo, Box<Error>> { ... }
    fn apply(&self, input: AlgoIo) -> Result<AlgoIo, Box<Error>> { ... }
}

Implementing an algorithm involves overriding at least one of these methods

Provided Methods

fn apply_str(&self, name: &str) -> Result<AlgoIo, Box<Error>>

fn apply_json(&self, json: &Json) -> Result<AlgoIo, Box<Error>>

fn apply_bytes(&self, bytes: &[u8]) -> Result<AlgoIo, Box<Error>>

fn apply(&self, input: AlgoIo) -> Result<AlgoIo, Box<Error>>

The default implementation of this method calls apply_str, apply_json, or apply_bytes based on the input type.

AlgoIo::Text results in call to apply_str

AlgoIo::Json results in call to apply_json

AlgoIo::Binary results in call to apply_bytes

If that call returns an UnsupportedInput error, then this method method will may attempt to coerce the input into another type and attempt one more call:

AlgoIo::Text input will be JSON-encoded to call apply_json

AlgoIo::Json input will be parse to see it can call apply_str