Struct algorithmia::algorithm::Algorithm

pub struct Algorithm<'a> {
    pub service: Service,
    pub user: &'a str,
    pub repo: &'a str,
    pub version: Version<'a>,
}

Algorithmia algorithm

Fields

service

user

repo

version

Methods

impl<'a> Algorithm<'a>

fn to_url(&self) -> Url

Get the API Endpoint URL for a particular algorithm

fn pipe<D, E>(&'a self, input_data: &E) -> AlgorithmResult<D> where D: Decodable, E: Encodable

pipeute an algorithm with typed JSON response decoding

input_data must be JSON-encodable use #[derive(RustcEncodable)] for complex input

You must explicitly specify the output type T pipe will attempt to decode the response into AlgorithmOutput

If decoding fails, it will attempt to decode into ApiError and if that fails, it will error with DecoderErrorWithContext

Examples

let service = Service::new("111112222233333444445555566");
let factor_service = service.algorithm("kenny", "Factor", Version::Latest);
let input = "19635".to_string();
match factor_service.pipe(&input) {
    Ok(out) => {
        let myVal: AlgorithmOutput<Vec<i64>> = out;
        println!("{:?}", myVal.result);
    },
    Err(AlgorithmiaError::ApiError(error)) => {
        println!("API Error: {:?}", error)
    },
    Err(e) => println!("ERROR: {:?}", e),
};

fn pipe_raw(&'a self, input_data: &str) -> AlgorithmJsonResult

pipeute an algorithm with with string input and receive the raw JSON response

pipe provides an encoding/decoding wrapper around this method

TODO: Understand if we need to support NOT setting Content-Type to application/json if the input isn't actually JSON

Examples

let algo_service = Service::new("111112222233333444445555566");
let factor  = algo_service.algorithm("kenny", "Factor", Version::Latest);

let output = match factor.pipe_raw("37") {
   Ok(result) => result,
   Err(why) => panic!("{:?}", why),
};