Struct algorithmia::algorithm::AlgorithmService

pub struct AlgorithmService<'a> {
    pub service: Service,
    pub algorithm: Algorithm<'a>,
}

Service endpoint for executing algorithms

Fields

service

algorithm

Methods

impl<'c> AlgorithmService<'c>

fn new(api_key: &'c str, user: &'c str, repo: &'c str, version: Version<'c>) -> AlgorithmService<'c>

Instantiate AlgorithmService directly - alternative to Service::algorithm

Examples

let mut factor = AlgorithmService::new("111112222233333444445555566", "kenny", "Factor", Version::Latest);

fn exec<'a, D, E>(&'c self, input_data: &E) -> AlgorithmResult<D> where D: Decodable, E: Encodable

Execute an algorithm with type-safety input_data must be JSON-encodable use #[derive(RustcEncodable)] for complex input

You must explicitly specify the output type T exec will attempt to decode the response into AlgorithmOutput

If decoding fails, it will attempt to decode into ApiError and if that fails, it will error with DecoderErrorWithContext

Examples

let algo_service = Service::new("111112222233333444445555566");
let mut factor = algo_service.algorithm("kenny", "Factor", Version::Latest);
let input = "19635".to_string();
match factor.exec(&input) {
    Ok(out) => {
        let myVal: AlgorithmOutput<Vec<i64>> = out;
        println!("{:?}", myVal.result);
    },
    Err(AlgorithmiaError::ApiError(error)) => {
        println!("API Error: {:?}", error)
    },
    Err(e) => println!("ERROR: {:?}", e),
};

fn exec_raw(&'c self, input_data: &str) -> AlgorithmJsonResult

Execute an algorithm with with string input and receive the raw JSON response

exec provides an encoding/decoding wrapper around this method

TODO: Understand if we need to support NOT setting Content-Type to application/json if the input isn't actually JSON

Examples

let algo_service = Service::new("111112222233333444445555566");
let mut factor = algo_service.algorithm("kenny", "Factor", Version::Latest);

let output = match factor.exec_raw("37") {
   Ok(result) => result,
   Err(why) => panic!("{:?}", why),
};