Module algorithmia::client

Structs

HttpClient
Internal HttpClient to build requests: wraps hyper client

Response
A response for a client request to a remote server.

Enums

Body
An enum of possible body types for a Request.