Module algorithmia::data::file

File module for managing Algorithmia Data Files

Examples

use algorithmia::Algorithmia;

let client = Algorithmia::client("111112222233333444445555566");
let my_file = client.file(".my/my_dir/some_filename");

my_file.put("file_contents");

Structs

DataFile
Algorithmia data file

DataResponse

FileAdded
Response when creating a file via the Data API

FileDeleted
Response when deleting a file from the Data API