Module algorithmia::client

Internal client

Do not use directly - use the Algorithmia struct instead

Structs

HttpClient
Internal HttpClient to build requests: wraps hyper client

Response
A response for a client request to a remote server.

Enums

ApiAuth
Represent the different ways to auth with the API

Body
An enum of possible body types for a Request.